+++ python/helpers/tensortiler/__init__.py
@@ -1,5 +1,5 @@
-from .tensortile import TensorTile
-from .tensortilesequence import (
-    TensorTileSequence,
+from .tensoraccesspattern import TensorAccessPattern
+from .tensoraccesssequence import (
+    TensorAccessSequence,
 )
 from .tensortiler2d import TensorTiler2D
